Feat: Add seperate createNote and updateNote
Created seperate functions for create and update.
Put the fetchNote in a seperate function which now fetches the note to the titleField & textarea field as well as changing the global variable ID.
By clicking on the edit icon we change the dataMode to 'true' which then converts the save button to an update button.